\name{dist2d}
\alias{dist2d}
\title{Distance of two points on a plane}
\usage{
  dist2d(x0, y0, x1, y1)
}
\arguments{
  \item{x0}{x-coordinate of 1st point}

  \item{y0}{y-coordinate of 1st point}

  \item{x1}{x-coordinate of 2nd point}

  \item{y1}{y-coordinate of 2nd point}
}
\value{
  The distance between the two points.
}
\description{
  Function to compute the distance of two points on a plane
}
\examples{
dist2d(0,0,1,1)   # sqrt(2)
}
\author{
  David Kneis \email{david.kneis@uni-potsdam.de}
}

